Restructuring ckeditor5-list package (splitting into subdirectories).
