+++ packages/ckeditor5-list/src/list.js
@@ -7,16 +7,16 @@
  * @module list/list
  */
 
-import ListEditing from './listediting';
-import ListUI from './listui';
+import ListEditing from './list/listediting';
+import ListUI from './list/listui';
 
 import { Plugin } from 'ckeditor5/src/core';
 
 /**
  * The list feature.
  *
- * This is a "glue" plugin that loads the {@link module:list/listediting~ListEditing list editing feature}
- * and {@link module:list/listui~ListUI list UI feature}.
+ * This is a "glue" plugin that loads the {@link module:list/list/listediting~ListEditing list editing feature}
+ * and {@link module:list/list/listui~ListUI list UI feature}.
  *
  * @extends module:core/plugin~Plugin
  */
